requirements_quantum.txt
absl-py==2.1.0
appdirs==1.4.4
argcomplete==3.5.3
astunparse==1.6.3
autograd==1.7.0
autoray==0.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
flatbuffers==24.12.23
fonttools==4.55.3
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.69.0
h5py==3.12.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
Jinja2==3.1.6
joblib==1.4.2
keras==2.14.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
ml-dtypes==0.2.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.0
nvidia-nccl-cu12==2.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
PennyLane==0.39.0
PennyLane_Lightning==0.39.0
pillow==11.1.0
pip==22.0.2
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==3.0.3
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
rustworkx==0.15.1
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.13.2
setuptools==59.6.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
statsmodels==0.14.4
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
umap==0.1.1
umap-learn==0.5.7
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.1
xgboost==2.1.3